Rename components from publish to deploy
