+++ web/src/components/DeployProgressSummary.vue
