+++ web/src/router.ts
@@ -6,7 +6,7 @@ import ProjectPage from 'src/views/project-page/ProjectPage.vue';
 import AddNewDeployment from 'src/views/add-new-deployment/AddNewDeployment.vue';
 import ExistingDeploymentPage from 'src/views/existing-deployment/ExistingDeploymentPage.vue';
 import NewDeploymentPage from 'src/views/new-deployment/NewDeploymentPage.vue';
-import PublishProgressPage from 'src/views/publish-progress/PublishProgressPage.vue';
+import DeployProgressPage from 'src/views/deploy-progress/DeployProgressPage.vue';
 import FatalErrorPage from 'src/views/fatal-error/FatalErrorPage.vue';
 
 const routes = [
@@ -22,7 +22,7 @@ const routes = [
     }),
   },
   { name: 'deployments', path: '/deployments/:name', component: ExistingDeploymentPage },
-  { name: 'progress', path: '/progress', component: PublishProgressPage },
+  { name: 'progress', path: '/progress', component: DeployProgressPage },
   {
     name: 'fatalError',
     path: '/error',
