+++ web/src/views/deploy-progress/DeployStep.vue
