+++ web/src/views/deploy-progress/steps/CreateBundle.vue
@@ -1,11 +1,11 @@
 <!-- Copyright (C) 2023 by Posit Software, PBC. -->
 
 <template>
-  <PublishStep
+  <DeployStep
     :name="name"
-    title="Run Content"
-    icon="sync"
-    summary="Performing execution checks ahead of applying settings."
+    title="Create Bundle"
+    icon="compress"
+    summary="Collecting and bundling up the files included in your project, so that they can be uploaded to the server within a bundle."
     :done="done"
     :messages="messages"
   />
@@ -14,7 +14,7 @@
 <script setup lang="ts">
 import { ref } from 'vue';
 
-import PublishStep from 'src/views/publish-progress/PublishStep.vue';
+import DeployStep from 'src/views/deploy-progress/DeployStep.vue';
 
 import { watch } from 'vue';
 import { useEventStore } from 'src/stores/events';
@@ -27,10 +27,10 @@ defineProps({
 const emit = defineEmits(['start', 'done']);
 
 const done = ref(false);
-const messages = ref(eventStore.currentPublishStatus.status.steps.runContent.allMsgs);
+const messages = ref(eventStore.currentPublishStatus.status.steps.createBundle.allMsgs);
 
 watch(
-  () => eventStore.currentPublishStatus.status.steps.runContent.completion,
+  () => eventStore.currentPublishStatus.status.steps.createBundle.completion,
   (value) => {
     if (value === 'inProgress') {
       emit('start');
@@ -43,5 +43,6 @@ watch(
     immediate: true,
   }
 );
+
 </script>
