+++ web/src/views/deploy-progress/steps/CreateDeployment.vue
@@ -1,11 +1,11 @@
 <!-- Copyright (C) 2023 by Posit Software, PBC. -->
 
 <template>
-  <PublishStep
+  <DeployStep
     :name="name"
-    title="Validate Deployment"
-    icon="sync"
-    summary="Validating your content on the Connect Server"
+    title="Create Deployment"
+    icon="create_new_folder"
+    summary="Registering the deployment object with the Posit Connect Server."
     :done="done"
     :messages="messages"
   />
@@ -14,7 +14,7 @@
 <script setup lang="ts">
 import { ref } from 'vue';
 
-import PublishStep from 'src/views/publish-progress/PublishStep.vue';
+import DeployStep from 'src/views/deploy-progress/DeployStep.vue';
 
 import { watch } from 'vue';
 import { useEventStore } from 'src/stores/events';
@@ -27,10 +27,10 @@ defineProps({
 const emit = defineEmits(['start', 'done']);
 
 const done = ref(false);
-const messages = ref(eventStore.currentPublishStatus.status.steps.validateDeployment.allMsgs);
+const messages = ref(eventStore.currentPublishStatus.status.steps.createDeployment.allMsgs);
 
 watch(
-  () => eventStore.currentPublishStatus.status.steps.validateDeployment.completion,
+  () => eventStore.currentPublishStatus.status.steps.createDeployment.completion,
   (value) => {
     if (value === 'inProgress') {
       emit('start');
@@ -44,4 +44,3 @@ watch(
   }
 );
 </script>
-
