+++ web/src/views/deploy-progress/steps/RestorePythonEnvironment.vue
@@ -1,11 +1,11 @@
 <!-- Copyright (C) 2023 by Posit Software, PBC. -->
 
 <template>
-  <PublishStep
+  <DeployStep
     :name="name"
-    title="Create Bundle"
-    icon="compress"
-    summary="Collecting and bundling up the files included in your project, so that they can be uploaded to the server within a bundle."
+    title="Restore Python Environment"
+    icon="move_down"
+    summary="Installing the dependent python packages on the server in order to reproduce your runtime environment."
     :done="done"
     :messages="messages"
   />
@@ -14,23 +14,22 @@
 <script setup lang="ts">
 import { ref } from 'vue';
 
-import PublishStep from 'src/views/publish-progress/PublishStep.vue';
+import DeployStep from 'src/views/deploy-progress/DeployStep.vue';
 
 import { watch } from 'vue';
 import { useEventStore } from 'src/stores/events';
 
 const eventStore = useEventStore();
-
 defineProps({
   name: { type: [String, Number], required: true },
 });
 const emit = defineEmits(['start', 'done']);
 
 const done = ref(false);
-const messages = ref(eventStore.currentPublishStatus.status.steps.createBundle.allMsgs);
+const messages = ref(eventStore.currentPublishStatus.status.steps.restorePythonEnv.allMsgs);
 
 watch(
-  () => eventStore.currentPublishStatus.status.steps.createBundle.completion,
+  () => eventStore.currentPublishStatus.status.steps.restorePythonEnv.completion,
   (value) => {
     if (value === 'inProgress') {
       emit('start');
@@ -43,6 +42,5 @@ watch(
     immediate: true,
   }
 );
-
 </script>
