+++ web/src/views/deploy-progress/steps/RunContent.vue
@@ -1,11 +1,11 @@
 <!-- Copyright (C) 2023 by Posit Software, PBC. -->
 
 <template>
-  <PublishStep
+  <DeployStep
     :name="name"
-    title="Upload Bundle"
-    icon="login"
-    summary="Transferring the files from your local workstation to the server."
+    title="Run Content"
+    icon="sync"
+    summary="Performing execution checks ahead of applying settings."
     :done="done"
     :messages="messages"
   />
@@ -14,22 +14,23 @@
 <script setup lang="ts">
 import { ref } from 'vue';
 
-import PublishStep from 'src/views/publish-progress/PublishStep.vue';
+import DeployStep from 'src/views/deploy-progress/DeployStep.vue';
 
 import { watch } from 'vue';
 import { useEventStore } from 'src/stores/events';
 
 const eventStore = useEventStore();
+
 defineProps({
   name: { type: [String, Number], required: true },
 });
 const emit = defineEmits(['start', 'done']);
 
 const done = ref(false);
-const messages = ref(eventStore.currentPublishStatus.status.steps.uploadBundle.allMsgs);
+const messages = ref(eventStore.currentPublishStatus.status.steps.runContent.allMsgs);
 
 watch(
-  () => eventStore.currentPublishStatus.status.steps.uploadBundle.completion,
+  () => eventStore.currentPublishStatus.status.steps.runContent.completion,
   (value) => {
     if (value === 'inProgress') {
       emit('start');
