+++ web/src/views/deploy-progress/steps/index.ts
